refactor: 把VisitRecordDataProc项目改成专门的数据处理项目，移动到 tools 目录下
