+++ StarBlog.sln
@@ -18,8 +18,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParallelTest", "demo\Parall
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenerateRandomCoordinates", "demo\GenerateRandomCoordinates\GenerateRandomCoordinates.csproj", "{CCA99BCE-D5A9-4A2E-8578-CCCE67F0C743}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VisitRecordDataProc", "demo\VisitRecordDataProc\VisitRecordDataProc.csproj", "{CCB07064-A3FC-446B-876E-ADF8D972B867}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{2FA94B80-AA27-4B92-9313-FB0311BCCEBD}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tools", "tools", "{BA52346E-15B5-46B9-8375-BEACDA6E3C04}"
@@ -36,6 +34,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StarBlog.Share", "src\StarB
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StarBlog.Web", "src\StarBlog.Web\StarBlog.Web.csproj", "{8CE57AC4-FBB2-4B36-8269-EA67D61A3037}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataProc", "tools\DataProc\DataProc.csproj", "{F2AB9CC8-E25E-4EC9-8B10-A506859BED74}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -47,13 +47,13 @@ Global
 		{70FCE492-23CE-44D0-AE8B-E07D8F309EB6} = {68E4117F-1EC3-4B1D-B413-888607595781}
 		{EB688698-734F-4169-99C5-8031B277FEB1} = {68E4117F-1EC3-4B1D-B413-888607595781}
 		{CCA99BCE-D5A9-4A2E-8578-CCCE67F0C743} = {68E4117F-1EC3-4B1D-B413-888607595781}
-		{CCB07064-A3FC-446B-876E-ADF8D972B867} = {68E4117F-1EC3-4B1D-B413-888607595781}
 		{4DD225BE-BAD1-4A68-B690-1493F352E400} = {BA52346E-15B5-46B9-8375-BEACDA6E3C04}
 		{4337439B-9B3C-40A8-B73B-B32842DEDD25} = {2FA94B80-AA27-4B92-9313-FB0311BCCEBD}
 		{C6AD7910-E5E5-4389-B471-7424B7DD7D02} = {2FA94B80-AA27-4B92-9313-FB0311BCCEBD}
 		{95C48311-A107-4E8E-B17E-F396474C1977} = {2FA94B80-AA27-4B92-9313-FB0311BCCEBD}
 		{6EDFB6BC-D41D-4772-A3DC-19DF6CFD088D} = {2FA94B80-AA27-4B92-9313-FB0311BCCEBD}
 		{8CE57AC4-FBB2-4B36-8269-EA67D61A3037} = {2FA94B80-AA27-4B92-9313-FB0311BCCEBD}
+		{F2AB9CC8-E25E-4EC9-8B10-A506859BED74} = {BA52346E-15B5-46B9-8375-BEACDA6E3C04}
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{0512E31F-954B-4E53-B3AF-F2EF206FAE33}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -76,10 +76,6 @@ Global
 		{CCA99BCE-D5A9-4A2E-8578-CCCE67F0C743}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CCA99BCE-D5A9-4A2E-8578-CCCE67F0C743}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CCA99BCE-D5A9-4A2E-8578-CCCE67F0C743}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CCB07064-A3FC-446B-876E-ADF8D972B867}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CCB07064-A3FC-446B-876E-ADF8D972B867}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CCB07064-A3FC-446B-876E-ADF8D972B867}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CCB07064-A3FC-446B-876E-ADF8D972B867}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4DD225BE-BAD1-4A68-B690-1493F352E400}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4DD225BE-BAD1-4A68-B690-1493F352E400}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4DD225BE-BAD1-4A68-B690-1493F352E400}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -104,5 +100,9 @@ Global
 		{8CE57AC4-FBB2-4B36-8269-EA67D61A3037}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8CE57AC4-FBB2-4B36-8269-EA67D61A3037}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8CE57AC4-FBB2-4B36-8269-EA67D61A3037}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F2AB9CC8-E25E-4EC9-8B10-A506859BED74}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F2AB9CC8-E25E-4EC9-8B10-A506859BED74}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F2AB9CC8-E25E-4EC9-8B10-A506859BED74}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F2AB9CC8-E25E-4EC9-8B10-A506859BED74}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
