--- demo/VisitRecordDataProc/src/Data/Config/AuditLogConfig.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using VisitRecordDataProc.Data.Models;
-
-namespace VisitRecordDataProc.Data.Config;
-
-public class AuditLogConfig : IEntityTypeConfiguration<AuditLog> {
-    public void Configure(EntityTypeBuilder<AuditLog> builder) {
-        builder.HasKey(x => x.Id);
-        builder.Property(x => x.Id).ValueGeneratedOnAdd();
-        builder.Property(x => x.EventType).IsRequired().HasMaxLength(256);
-        builder.Property(x => x.Username).IsRequired().HasMaxLength(256);
-        builder.Property(x => x.Timestamp).IsRequired();
-        builder.Property(x => x.EntityName).HasMaxLength(256);
-        builder.Property(x => x.EntityId).HasMaxLength(256);
-    }
-}
