--- demo/VisitRecordDataProc/src/Entities/AppSettings.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace VisitRecordDataProc.Entities;
-
-public class AppSettings {
-    public string Name { get; set; }
-    public bool Boolean { get; set; }
-    public List<string> DemoList { get; set; } = new();
-}
