--- demo/VisitRecordDataProc/src/Entities/OutputResult.cs
@@ -1,8 +1,0 @@
-using System.Collections.Generic;
-
-namespace VisitRecordDataProc.Entities;
-
-public class OutputResult {
-    public string Result { get; set; }
-    public IEnumerable<string> Messages { get; set; }
-}
