--- demo/VisitRecordDataProc/src/Framework/FluentConsoleBuilder.cs
@@ -1,20 +1,0 @@
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace VisitRecordDataProc.Framework;
-
-public class FluentConsoleBuilder {
-    public IConfiguration Configuration { get; set; }
-    public IServiceCollection Services { get; set; }
-
-    internal FluentConsoleBuilder() { }
-
-    public FluentConsoleApp Build() {
-        var app = new FluentConsoleApp {
-            Configuration = Configuration,
-            Services = Services
-        };
-
-        return app;
-    }
-}
