--- demo/VisitRecordDataProc/src/Properties/UserAgentInfoProfile.cs
@@ -1,13 +1,0 @@
-﻿using AutoMapper;
-using StarBlog.Data.Models;
-
-namespace VisitRecordDataProc.Properties;
-
-public class UserAgentInfoProfile : Profile {
-    public UserAgentInfoProfile() {
-        CreateMap<UAParser.OS, OS>();
-        CreateMap<UAParser.Device, Device>();
-        CreateMap<UAParser.UserAgent, UserAgent>();
-        CreateMap<UAParser.ClientInfo, UserAgentInfo>();
-    }
-}
