--- demo/VisitRecordDataProc/src/Services/IService.cs
@@ -1,7 +1,0 @@
-using FluentResults;
-
-namespace VisitRecordDataProc.Services;
-
-public interface IService {
-    Task<Result> Run();
-}
