+++ tools/DataProc/appsettings.json
@@ -1,0 +1,26 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.EntityFrameworkCore.Database.Command": "Warning",
+      "Microsoft.EntityFrameworkCore": "Information"
+    }
+  },
+  "ConnectionStrings": {
+    "SQLite": "Data Source=app.db;",
+    "SQLite-Log": "Data Source=C:\\code\\StarBlog\\StarBlog.Web\\app.log.db;",
+    "Ip2Region": "C:\\code\\StarBlog\\StarBlog.Web\\ip2region.xdb"
+  },
+  "StarBlog": {
+    "wwwroot": "C:\\code\\starblog\\starblog\\src\\StarBlog.Web\\wwwroot"
+  },
+  "AppSettings": {
+    "Name": "The name of this app is fluent console demo",
+    "Boolean": true,
+    "DemoList": [
+      "item1",
+      "item2",
+      "item3"
+    ]
+  }
+}
