+++ tools/DataProc/src/Data/AppDbContext.cs
@@ -1,0 +1,15 @@
+﻿using DataProc.Data.Models;
+using Microsoft.EntityFrameworkCore;
+
+namespace DataProc.Data;
+
+public class AppDbContext : DbContext {
+    public DbSet<AuditLog> AuditLogs { get; set; }
+
+    public AppDbContext(DbContextOptions<AppDbContext> options) : base(options) { }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder) {
+        base.OnConfiguring(optionsBuilder);
+        optionsBuilder.UseSnakeCaseNamingConvention();
+    }
+}
