+++ tools/DataProc/src/Entities/AppSettings.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace DataProc.Entities;
+
+public class AppSettings {
+    public string Name { get; set; }
+    public bool Boolean { get; set; }
+    public List<string> DemoList { get; set; } = new();
+}
