+++ tools/DataProc/src/Framework/FluentConsoleBuilder.cs
@@ -1,0 +1,20 @@
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace DataProc.Framework;
+
+public class FluentConsoleBuilder {
+    public IConfiguration Configuration { get; set; }
+    public IServiceCollection Services { get; set; }
+
+    internal FluentConsoleBuilder() { }
+
+    public FluentConsoleApp Build() {
+        var app = new FluentConsoleApp {
+            Configuration = Configuration,
+            Services = Services
+        };
+
+        return app;
+    }
+}
