+++ tools/DataProc/src/Services/IService.cs
@@ -1,0 +1,7 @@
+using FluentResults;
+
+namespace DataProc.Services;
+
+public interface IService {
+    Task<Result> Run();
+}
