+++ tools/DataProc/src/Utilities/ConsoleTool.cs
@@ -1,0 +1,53 @@
+using System;
+using System.Text;
+
+namespace DataProc.Utilities;
+
+public static class ConsoleTool {
+    public static void PrintLogo() {
+        byte[] logo = {
+            0xD, 0xA, 0x20, 0x5F, 0x5F, 0x5F, 0x5F, 0x20, 0x20, 0x20, 0x20, 0x5F, 0x5F, 0x20, 0x20, 0x20, 0x20, 0x20,
+            0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x5F, 0x5F, 0x5F,
+            0x5F, 0x20, 0x20, 0x20, 0x20, 0x5F, 0x5F, 0x5F, 0xD, 0xA, 0x2F, 0x5C, 0x20, 0x20, 0x5F, 0x60, 0x5C, 0x20,
+            0x2F, 0x5C, 0x20, 0x5C, 0x5F, 0x5F, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+            0x20, 0x20, 0x20, 0x20, 0x2F, 0x5C, 0x20, 0x20, 0x5F, 0x60, 0x5C, 0x20, 0x2F, 0x5C, 0x5F, 0x20, 0x5C, 0xD,
+            0xA, 0x5C, 0x20, 0x5C, 0x2C, 0x5C, 0x4C, 0x5C, 0x5F, 0x5C, 0x20, 0x5C, 0x20, 0x2C, 0x5F, 0x5C, 0x20, 0x20,
+            0x20, 0x20, 0x5F, 0x5F, 0x20, 0x20, 0x20, 0x20, 0x20, 0x5F, 0x20, 0x5F, 0x5F, 0x5C, 0x20, 0x5C, 0x20, 0x5C,
+            0x4C, 0x5C, 0x20, 0x5C, 0x2F, 0x2F, 0x5C, 0x20, 0x5C, 0x20, 0x20, 0x20, 0x20, 0x20, 0x5F, 0x5F, 0x5F, 0x20,
+            0x20, 0x20, 0x20, 0x20, 0x20, 0x5F, 0x5F, 0xD, 0xA, 0x20, 0x5C, 0x2F, 0x5F, 0x5C, 0x5F, 0x5F, 0x20, 0x5C,
+            0x5C, 0x20, 0x5C, 0x20, 0x5C, 0x2F, 0x20, 0x20, 0x2F, 0x27, 0x5F, 0x5F, 0x60, 0x5C, 0x20, 0x20, 0x2F, 0x5C,
+            0x60, 0x27, 0x5F, 0x5F, 0x5C, 0x20, 0x5C, 0x20, 0x20, 0x5F, 0x20, 0x3C, 0x27, 0x5C, 0x20, 0x5C, 0x20, 0x5C,
+            0x20, 0x20, 0x20, 0x2F, 0x20, 0x5F, 0x5F, 0x60, 0x5C, 0x20, 0x20, 0x2F, 0x27, 0x5F, 0x20, 0x60, 0x5C, 0xD,
+            0xA, 0x20, 0x20, 0x20, 0x2F, 0x5C, 0x20, 0x5C, 0x4C, 0x5C, 0x20, 0x5C, 0x20, 0x5C, 0x20, 0x5C, 0x5F, 0x2F,
+            0x5C, 0x20, 0x5C, 0x4C, 0x5C, 0x2E, 0x5C, 0x5F, 0x5C, 0x20, 0x5C, 0x20, 0x5C, 0x2F, 0x20, 0x5C, 0x20, 0x5C,
+            0x20, 0x5C, 0x4C, 0x5C, 0x20, 0x5C, 0x5C, 0x5F, 0x5C, 0x20, 0x5C, 0x5F, 0x2F, 0x5C, 0x20, 0x5C, 0x4C, 0x5C,
+            0x20, 0x5C, 0x2F, 0x5C, 0x20, 0x5C, 0x4C, 0x5C, 0x20, 0x5C, 0xD, 0xA, 0x20, 0x20, 0x20, 0x5C, 0x20, 0x60,
+            0x5C, 0x5F, 0x5F, 0x5F, 0x5F, 0x5C, 0x20, 0x5C, 0x5F, 0x5F, 0x5C, 0x20, 0x5C, 0x5F, 0x5F, 0x2F, 0x2E, 0x5C,
+            0x5F, 0x5C, 0x5C, 0x20, 0x5C, 0x5F, 0x5C, 0x20, 0x20, 0x5C, 0x20, 0x5C, 0x5F, 0x5F, 0x5F, 0x5F, 0x2F, 0x2F,
+            0x5C, 0x5F, 0x5F, 0x5F, 0x5F, 0x5C, 0x20, 0x5C, 0x5F, 0x5F, 0x5F, 0x5F, 0x2F, 0x5C, 0x20, 0x5C, 0x5F, 0x5F,
+            0x5F, 0x5F, 0x20, 0x5C, 0xD, 0xA, 0x20, 0x20, 0x20, 0x20, 0x5C, 0x2F, 0x5F, 0x5F, 0x5F, 0x5F, 0x5F, 0x2F,
+            0x5C, 0x2F, 0x5F, 0x5F, 0x2F, 0x5C, 0x2F, 0x5F, 0x5F, 0x2F, 0x5C, 0x2F, 0x5F, 0x2F, 0x20, 0x5C, 0x2F, 0x5F,
+            0x2F, 0x20, 0x20, 0x20, 0x5C, 0x2F, 0x5F, 0x5F, 0x5F, 0x2F, 0x20, 0x5C, 0x2F, 0x5F, 0x5F, 0x5F, 0x5F, 0x2F,
+            0x5C, 0x2F, 0x5F, 0x5F, 0x5F, 0x2F, 0x20, 0x20, 0x5C, 0x2F, 0x5F, 0x5F, 0x5F, 0x4C, 0x5C, 0x20, 0x5C, 0xD,
+            0xA, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+            0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+            0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+            0x20, 0x20, 0x20, 0x20, 0x20, 0x2F, 0x5C, 0x5F, 0x5F, 0x5F, 0x5F, 0x2F, 0xD, 0xA, 0x20, 0x20, 0x20, 0x20,
+            0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+            0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+            0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+            0x5C, 0x5F, 0x2F, 0x5F, 0x5F, 0x2F, 0xD, 0xA
+        };
+        Console.WriteLine(Encoding.ASCII.GetString(logo));
+    }
+
+    public static void PrintDivider() {
+        Console.WriteLine("============================================================================");
+    }
+
+    public static void PrintTitle(string title) {
+        PrintDivider();
+        Console.WriteLine(title);
+        PrintDivider();
+    }
+}
