+++ tools/DataProc/src/Utilities/SourceGenerationContext.cs
@@ -1,0 +1,10 @@
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using System.Text.Json.Serialization.Metadata;
+using DataProc.Entities;
+
+namespace DataProc.Utilities;
+
+[JsonSourceGenerationOptions(WriteIndented = true)]
+[JsonSerializable(typeof(OutputResult))]
+internal partial class SourceGenerationContext : JsonSerializerContext { }
